Section 테스트 코드 작성 및 로직 변경
1. Section 테스트 코드 작성

2. 로직 변경

- content 변경 -> sectionId body -> path variable로 변경

- orderIdx -> short에서 int 형으로 변경 (공간 절약 보다 생산성을 선택)

- 마지막 orderIdx 가져오기 -> Section이 없다면 0 대신 Optional.empty로 반환

orderIdx는 1번이 최소이기 때문에 변경

+++ src/main/java/seungyong/helpmebackend/section/adapter/in/web/dto/response/ResponseSections.java
@@ -36,6 +36,6 @@ public record ResponseSections(
             @Schema(description = "섹션 내용 (Markdown 형식)", example = "이 섹션은 프로젝트의 개요를 설명합니다.")
             String content,
             @Schema(description = "섹션 순서 (1부터 시작)", example = "1")
-            Short orderIdx
+            Integer orderIdx
     ) { }
 }
